Add files via upload
 Data-driven scenarios
I am testing integration between an eCommerce platform (does not matter which one but can imagine it is Magento) and ERP. There are two scenarios I am testing:

1) The first scenario is an update of the product price and quantity from the ERP system. ERP is sending to the Magento store system a file (CSV format) with information about updates on product price and product quantity. Magento consumes this file and makes an update for product price and stock value on the front end. 

2) The second scenario is an update of order status from ERP to Magento. Magento just receives an order and then sends it to ERP for order processing. Once there is a status change in ERP e.g. an order is paid, shipped, canceled, refunded, completed, or anything in between, ERP sends a file with the list of all orders that have order status change during a period of time, say one hour. 

Once Magento gets this file, it consumes it updates the order status in the customer account, and sends an update email. 

The problem is that the developer says that he finished the task, but ... the ERP system went offline and did not send any files to you! Ah... and the demo to the customer is just in a few hours! 

The developer says that he still can demo this integration and both scenarios if you would provide him with sample files from ERP. Basically, he asked me to give him one file that simulates ERP output for product price and stock update. It can be one file for 5 products. 

For the second scenario, he asks for 3-4 sample files, each having order status updates for a maximum of 5 orders. These files are generated every hour, say 12:00, 13:00, 14:00 and 15:00. 

Then he will consume this data and make changes in products and order statuses.